requirements.txt
absl-py==2.0.0
ale-py==0.8.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==2.2.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.10
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.13.1
flloat==0.3.0
fonttools==4.44.0
fsspec==2023.10.0
glfw==2.7.0
graphviz==0.20.1
gym==0.26.2
gym-notices==0.0.8
gym-super-mario-bros==7.4.0
gymnasium==0.29.1
idna==3.4
imageio==2.34.0
importlib_resources==6.5.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lark-parser==0.12.0
ltlf2dfa==1.0.2
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
mujoco-py==2.1.2.14
nes-py==8.2.1
networkx==3.2.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pycparser==2.22
pygame==2.5.2
pyglet==1.5.21
pyparsing==3.1.1
pythomata==0.3.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
